fix: domainServiceにタグの生成も含める
