+++ domain/root-post/rootPostDomainService.ts
@@ -1,3 +1,4 @@
+import { randomId } from "../../id";
 import RootPostRepoImpl from "../../infra/rootPostRepoImpl"
 import RootTagRepoImpl from "../../infra/rootTagRepoImpl"
 import { RootPost } from "./rootPost"
@@ -15,7 +16,37 @@ export class RootPostDomainService {
     this.teamRepo = teamRepo
   }
 
-  public async createRootPost(rootPost: RootPost, teamId: string) {
+  private addPostToRootTags(contents: string[], newPostId: string) {
+    let rootTags: RootTag[] = []
+
+    for (const content of contents) {
+      let newRootTag
+      const rootTag = this.tagRepo.findTagByContent(content)
+
+      if (rootTag) {
+        newRootTag = this.addPostToExistingRootTag(rootTag, newPostId)
+      } else {
+        newRootTag = this.createNewRootTag(content, newPostId)
+      }
+      rootTags.push(newRootTag)
+    }
+
+    return rootTags
+  }
+
+  private addPostToExistingRootTag(rootTag: RootTag, newPostId: string) {
+    return new RootTag(rootTag.tag.content, [...rootTag.postIds, newPostId])
+  }
+  private createNewRootTag(newTagContent:string, newPostId: string) {
+    return new RootTag(newTagContent, [newPostId])
+  }
+
+  public async createRootPost(content: string, tagContents: string[], teamId: string) {
+    const newPostId = randomId()
+    const rootTags = this.addPostToRootTags(tagContents, newPostId)
+    const rootPost = new RootPost({id: newPostId, content, tagIds: rootTags.map((rootTag) => rootTag.tag.id)})
+    await this.tagRepo.saveAll(rootTags)
+
     const team = await this.teamRepo.find(teamId)
     if (!team.isPaid()) {
       throw Error("This team has not paid!")
